+++ deploy/sql/dummy-data/school.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const sql = require('mssql')
-const config = require('../../../infra/config')
+const config = require('../config')
 
 const schoolCount = 20000
